Added a number of small things (input processing), as well as some things to do with PPPM Also changed some things to do with pressure coupling. The pressure is now in bar everywhere, so that constants to do with Pcoupling should also use that unit (reference pressure, compressibility). The infamous presfac which converts from MD units to bar, now only occurs on two places: at the calculation of pressure from the virial, and for calculation of long range corrections to the pressure.

+++ src/gmxlib/gmxfio.c
@@ -388,7 +388,7 @@ static bool do_binread(void *item,int nitem,int eio,
     rsize = fread(item,size,nitem,curfio->fp);
   else {
     /* Skip over it if we have a NULL pointer here */
-    fseek(curfio->fp,size*nitem,SEEK_CUR);
+    fseek(curfio->fp,(long) (size*nitem),SEEK_CUR);
     rsize = nitem;
   }
   if ((rsize != nitem) && (curfio->bDebug))
